examples/python/services/voice/pyproject.toml
[build-system]
requires = [
  "setuptools>=61.0",
  "wheel"
]
build-backend = "setuptools.build_meta"

[project]
version = "0.0.11"
name = "sample-browserbase"
description = "Browserbase Chatbot"
requires-python = ">=3.9.0"
license = "Apache-2.0"
classifiers = []
keywords = []
dependencies = [
  "meshagent-api~=0.0.16",
  "meshagent-agents~=0.0.16",
  "meshagent-openai~=0.0.16",
  "meshagent-livekit~=0.0.16",
]

[tool.setuptools.package-data]
"meshagent.agents" = ["py.typed", "*.pyi", "**/*.pyi", "**/*.js"]